restructure src into packages
