+++ .pre-commit-config.yaml
@@ -3,6 +3,7 @@ repos:
     rev: 4.0.1
     hooks:
       - id: flake8
+        args: ["--ignore=F401,W503"]
   - repo: https://github.com/kynan/nbstripout
     rev: 0.5.0
     hooks:
