+++ config/refresh_source_capta/all_parks.yaml
@@ -3,7 +3,7 @@ ACAD: Acadia NP
 ADAM: Adams NHP
 AFBG: African Burial Ground NM
 AGFO: Agate Fossil Beds NM
-ALAG: Alagnak Wild River
+ALAG: Alagnak W&SR
 ALFL: Alibates Flint Quarries NM
 ALPO: Allegheny Portage Railroad NHS
 AMIS: Amistad NRA
@@ -39,7 +39,7 @@ BOWA: Booker T. Washington NM
 BOAF: Boston African American NHS
 BOHA: Boston Harbor Islands NRA
 BOST: Boston NHP
-BRCR: Brices Cross Roads
+BRCR: Brices Cross Roads NBS
 BRVB: Brown v. Board of Education NHS
 BRCA: Bryce Canyon NP
 BUIS: Buck Island Reef NM
@@ -199,7 +199,7 @@ JECA: Jewel Cave NM
 JICA: Jimmy Carter NHP
 JODR: John D. Rockefeller, Jr. MEM PKWY
 JODA: John Day Fossil Beds NM
-JOFK: John F. Kennedy Center For Pa
+JOFK: John F. Kennedy Center For Performing Arts
 JOFI: John F. Kennedy NHS
 JOMU: John Muir NHS
 JOFL: Johnstown Flood NMEM
@@ -325,13 +325,13 @@ ROLA: Ross Lake NRA
 RUCA: Russell Cave NM
 SAHI: Sagamore Hill NHS
 SAGU: Saguaro NP
-SACR: Saint Croix Island IHS
+SACR: Saint Croix Island NHS
 SACN: Saint Croix NSR
 SAPA: Saint Paul's Church NHS
 SAGA: Saint-Gaudens NHP
 SAMA: Salem Maritime NHS
 SAPU: Salinas Pueblo Missions NM
-SARI: Salt River Bay NHP & Ecological Pres
+SARI: Salt River Bay NHP & Ecological Preserve
 SAAN: San Antonio Missions NHP
 SAFR: San Francisco Maritime NHP
 SAJH: San Juan Island NHP
@@ -373,7 +373,7 @@ TUAI: Tuskegee Airmen NHS
 TUIN: Tuskegee Institute NHS
 TUZI: Tuzigoot NM
 ULSG: Ulysses S. Grant NHS
-UPDE: Upper Delaware S&RR
+UPDE: Upper Delaware W&SR
 VALL: Valles Caldera NPRES
 VAFO: Valley Forge NHP
 VAMA: Vanderbilt Mansion NHS
