+++ config/refresh_source_capta/park_types.yaml
@@ -1,3 +1,5 @@
+EHP: Ecological and Historic Preserve
+MEM: Memorial
 NB: National Battlefield
 NBS: National Battlefield Site
 NHP: National Historical Park
@@ -15,5 +17,6 @@ NRRA: National River and Recreation Area
 NS : National Seashore
 NSR: National Scenic River
 NST: National Scenic Trail
+PK: Park
 PKWY: Parkway
-WSR: Wild and Scenic River
+"W&SR": Wild and Scenic River
