+++ config/train_models/recipes.yaml
@@ -11,9 +11,9 @@
 #     is highly recommended for organizational purposes
 #   * algorithm: the name of an ML modeling process to be applied, which should
 #     be one of the options listed in the _allowable_algorithms object in the
-#     __init__() method of the src.national_park_model.NationalParkModel() class
+#     __init__() method of the src.national_park_model.NationalParksModel() class
 #   * params (optional): any keyword arguments required by the indicated ML
-#     algorithm, for which see the implementations in src/national_park_model.py
+#     algorithm, for which see the implementations in src/national_parks_model.py
 
 - name: SARIMAXs for Individual Parks
   input: visitors_by_park
