+++ dvc.lock
@@ -43,18 +43,16 @@ stages:
     cmd: python src/refresh_source_capta.py
     deps:
     - path: config/refresh_source_capta
-      md5: 0f0de40904b24e11c79766c93b21166b.dir
-      size: 11174
+      md5: 3979f3b3f9e079987e326bd570a193d2.dir
+      size: 11253
       nfiles: 3
-    - path: src/np_scraper.py
-      md5: 71c4ce01ef91ccbb864225483214c7f7
-      size: 5587
-    - path: src/np_source_captaset.py
-      md5: fe9999d59ec70bfa838506fc8c9af0f2
-      size: 2485
+    - path: src/national_parks/nps
+      md5: d794ce25888521130cfa07a5c963cf4e.dir
+      size: 16747
+      nfiles: 6
     - path: src/refresh_source_capta.py
-      md5: 55467026d833007e43b0935a8cdf0107
-      size: 1952
+      md5: cc28db7fdef5cd75934ccff57ffe3e36
+      size: 1956
     outs:
     - path: capta/source
       md5: 6917197345dde3c4424ae52b3f244580.dir
@@ -68,15 +66,16 @@ stages:
       size: 23993
       nfiles: 1
     - path: config/process_capta
-      md5: d751713988987e9331980363e24189ce.dir
-      size: 0
-      nfiles: 0
+      md5: 7a9c5bbf919d20037ab41145cfd7a257.dir
+      size: 3101
+      nfiles: 2
+    - path: src/national_parks/processing
+      md5: 3881da2a17242da11b3cae827a9abd8f.dir
+      size: 8679
+      nfiles: 6
     - path: src/process_capta.py
-      md5: f5172cc976794e36a6f776d05ce12a23
-      size: 2050
-    - path: src/transformations.py
-      md5: dd6a92b4534158c5e75fd5ce1831a271
-      size: 3550
+      md5: 6839fae81d481bea488b44e154cae3f1
+      size: 2464
     outs:
     - path: capta/processed
       md5: 9fe5b0de295c51ab8de27fe74de98dda.dir
@@ -90,24 +89,26 @@ stages:
       size: 63238
       nfiles: 2
     - path: config/train_models
-      md5: d751713988987e9331980363e24189ce.dir
-      size: 0
-      nfiles: 0
-    - path: src/national_park_model.py
-      md5: 6af365e521d0f87a25e39dda89133751
-      size: 9008
+      md5: c8b18ad09192a6a5aaf27d905addfa0f.dir
+      size: 2259
+      nfiles: 2
+    - path: src/national_parks/model
+      md5: 951b2e3f1f8530b2a2d045e112d59d80.dir
+      size: 20836
+      nfiles: 6
+    - path: src/national_parks/visualization
+      md5: 694771ec474f6b040519e81f793239b2.dir
+      size: 18972
+      nfiles: 8
     - path: src/train_models.py
-      md5: d8ed0e217d1f75b84032db5c9502a09f
-      size: 2279
-    - path: src/visualization.py
-      md5: 5901377c35f5d62d44fa3e233a7996bb
-      size: 7158
+      md5: 731dd895cda8e5ce62a92f43c81a7cb3
+      size: 2528
     outs:
     - path: models
-      md5: cb6271daba39ba70e38385c38cb6984b.dir
+      md5: 16f3de00a5f2af95ef6eb59a5a2b5e20.dir
       size: 368852452
       nfiles: 26
     - path: plots
-      md5: 76c15daa7380fdd7375a8c3d15ab3f56.dir
-      size: 2362313
-      nfiles: 26
+      md5: ded97ba8eadc49c63e6c6551d7bebfe0.dir
+      size: 4086744
+      nfiles: 39
