+++ src/national_parks/__init__.py
@@ -1,0 +1,3 @@
+"""Package for sequestering functionality away from the DVC pipeline stage
+runner scripts.
+"""
