+++ src/national_parks/model/__init__.py
@@ -1,0 +1,5 @@
+"""Package to manage classes and operations associated with the train_models
+DVC stage.
+"""
+
+from .national_parks_model import NationalParksModel
