+++ src/national_parks/nps/park_scraper.py
@@ -43,13 +43,13 @@ def _bs_table_to_pandas(html_table, header=True):
     return df
 
 
-class NationalParkScraper(object):
+class NPSParkScraper(object):
     """Webscraping and capta storage class for an individual national park.
 
     Attributes:
-        name (str): the abbreviated name or 'code' for a park (e.g., 'ACAD')
+        name (str): the abbreviated name or 'code' for a park (e.g., "ACAD")
         park_type (str): the type of park (must be listed in
-            config/refresh_source_capta/park_types.yaml, e.g., 'NP')
+            config/refresh_source_capta/park_types.yaml, e.g., "NP")
         _monthly_visitors (pd.DataFrame): a DataFrame (NB: use
             get_monthly_visitors() to retrieve capta associated with this
             object, do not retrieve it directly)
