+++ src/national_parks/processing/__init__.py
@@ -1,0 +1,5 @@
+"""Package to manage classes and operations associated with the process_capta
+DVC stage.
+"""
+
+from .transformation import Transformation
