+++ src/national_parks/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Utilities package for the national-parks project."""
