+++ src/national_parks/visualization/__init__.py
@@ -1,0 +1,1 @@
+"""Package to manage the creation of plots and other visualizations."""
